# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Mercanto — Marketplace Vecinal API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
  message?: string;
}

export interface SuccessResponse {
  success: boolean;
  message?: string;
}

export interface RegisterRequest {
  name: string;
  email: string;
  /** @minLength 6 */
  password: string;
  phone?: string;
  district?: string;
}

export interface LoginRequest {
  email: string;
  password: string;
}

export interface UserPublic {
  id: number;
  name: string;
  email: string;
  role: string;
  avatarUrl?: string | null;
  phone?: string | null;
  district?: string | null;
  isBlocked: boolean;
  createdAt?: string | null;
}

export interface AuthResponse {
  user: UserPublic;
  token: string;
}

export interface UpdateProfileRequest {
  name?: string;
  phone?: string;
  district?: string;
  avatarUrl?: string;
}

export interface UserListResponse {
  users: UserPublic[];
  total: number;
  page: number;
  totalPages: number;
}

export interface Category {
  id: number;
  name: string;
  slug: string;
  icon?: string | null;
  description?: string | null;
  parentId?: number | null;
  isActive?: boolean | null;
  sortOrder?: number | null;
  createdAt?: string | null;
}

export interface CreateCategoryRequest {
  name: string;
  slug: string;
  icon?: string;
  description?: string;
  isActive?: boolean;
  sortOrder?: number;
}

export interface Store {
  id: number;
  userId: number;
  name: string;
  slug: string;
  description?: string | null;
  logoUrl?: string | null;
  bannerUrl?: string | null;
  categoryId?: number | null;
  location?: string | null;
  district?: string | null;
  lat?: string | null;
  lng?: string | null;
  whatsapp?: string | null;
  instagram?: string | null;
  facebook?: string | null;
  website?: string | null;
  status: string;
  isFeatured?: boolean | null;
  totalVisits?: number | null;
  createdAt?: string | null;
  updatedAt?: string | null;
  category?: Category | null;
}

export interface ProductImage {
  id: number;
  productId: number;
  url: string;
  publicId: string;
  sortOrder?: number | null;
}

export interface ProductWithImages {
  id: number;
  storeId: number;
  categoryId?: number | null;
  name: string;
  slug: string;
  description?: string | null;
  price: string;
  offerPrice?: string | null;
  stock?: number | null;
  unit?: string | null;
  status: string;
  sortOrder?: number | null;
  createdAt?: string | null;
  images?: ProductImage[];
  category?: Category | null;
}

export interface GalleryImage {
  id: number;
  storeId: number;
  url: string;
  publicId: string;
  sortOrder?: number | null;
}

export type StoreDetail = Store & {
  products?: ProductWithImages[];
  galleryImages?: GalleryImage[];
  reviewCount?: number;
  averageRating?: number | null;
};

export interface StoreMapItem {
  id: number;
  name: string;
  slug: string;
  description?: string | null;
  logoUrl?: string | null;
  district?: string | null;
  lat: string;
  lng: string;
  whatsapp?: string | null;
  category?: Category | null;
}

export interface StoreListResponse {
  stores: Store[];
  total: number;
  page: number;
  totalPages: number;
}

export interface CreateStoreRequest {
  name: string;
  description?: string;
  categoryId?: number;
  location?: string;
  district?: string;
  lat?: string;
  lng?: string;
  whatsapp?: string;
  instagram?: string;
  facebook?: string;
  website?: string;
  logoUrl?: string;
  logoPublicId?: string;
  bannerUrl?: string;
  bannerPublicId?: string;
}

export type UpdateStoreRequest = CreateStoreRequest;

export type UpdateStoreStatusRequestStatus =
  (typeof UpdateStoreStatusRequestStatus)[keyof typeof UpdateStoreStatusRequestStatus];

export const UpdateStoreStatusRequestStatus = {
  pending: "pending",
  active: "active",
  rejected: "rejected",
} as const;

export interface UpdateStoreStatusRequest {
  status: UpdateStoreStatusRequestStatus;
}

export type CreateProductRequestImagesItem = {
  url: string;
  publicId: string;
};

export interface CreateProductRequest {
  name: string;
  description?: string;
  price: string;
  offerPrice?: string;
  stock?: number;
  unit?: string;
  categoryId?: number;
  status?: string;
  images?: CreateProductRequestImagesItem[];
}

export interface ReviewWithUser {
  id: number;
  storeId: number;
  userId: number;
  rating: number;
  comment?: string | null;
  isVisible?: boolean | null;
  createdAt?: string | null;
  user?: UserPublic;
}

export interface CreateReviewRequest {
  /**
   * @minimum 1
   * @maximum 5
   */
  rating: number;
  comment?: string;
}

export interface SearchResults {
  stores: Store[];
  products: ProductWithImages[];
  query: string;
}

export interface Banner {
  id: number;
  title?: string | null;
  subtitle?: string | null;
  imageUrl: string;
  publicId?: string | null;
  linkUrl?: string | null;
  isActive?: boolean | null;
  sortOrder?: number | null;
}

export interface CreateBannerRequest {
  title?: string;
  subtitle?: string;
  imageUrl: string;
  publicId?: string;
  linkUrl?: string;
  isActive?: boolean;
  sortOrder?: number;
}

export interface AdminStats {
  totalUsers: number;
  totalStores: number;
  activeStores: number;
  pendingStores: number;
  totalProducts: number;
  totalReviews: number;
}

export type UpdateUserRoleRequestRole =
  (typeof UpdateUserRoleRequestRole)[keyof typeof UpdateUserRoleRequestRole];

export const UpdateUserRoleRequestRole = {
  user: "user",
  vendor: "vendor",
  admin: "admin",
} as const;

export interface UpdateUserRoleRequest {
  role: UpdateUserRoleRequestRole;
}

export type ListStoresParams = {
  district?: string;
  category?: string;
  search?: string;
  sort?: ListStoresSort;
  page?: number;
  limit?: number;
  featured?: boolean;
};

export type ListStoresSort =
  (typeof ListStoresSort)[keyof typeof ListStoresSort];

export const ListStoresSort = {
  newest: "newest",
  oldest: "oldest",
  name: "name",
  visits: "visits",
} as const;

export type ListStoresForMapParams = {
  district?: string;
  category?: string;
};

export type GetProductsByStoreParams = {
  category?: string;
};

export type GlobalSearchParams = {
  q: string;
  type?: GlobalSearchType;
  district?: string;
};

export type GlobalSearchType =
  (typeof GlobalSearchType)[keyof typeof GlobalSearchType];

export const GlobalSearchType = {
  all: "all",
  stores: "stores",
  products: "products",
} as const;

export type AdminListStoresParams = {
  status?: string;
  page?: number;
};

export type AdminListUsersParams = {
  page?: number;
};