# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Mercanto — Marketplace Vecinal API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /stores:
    get:
      operationId: listStores
      tags: [stores]
      summary: List active stores with filters
      parameters:
        - name: district
          in: query
          schema:
            type: string
        - name: category
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
        - name: sort
          in: query
          schema:
            type: string
            enum: [newest, oldest, name, visits]
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
        - name: featured
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: List of stores
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StoreListResponse"

    post:
      operationId: createStore
      tags: [stores]
      summary: Create a new store
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateStoreRequest"
      responses:
        "200":
          description: Store created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Store"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /admin/stores:
    get:
      operationId: adminListStores
      tags: [admin]
      summary: List all stores (admin)
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: All stores
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StoreListResponse"
  /admin/stores/{id}/status:
    put:
      operationId: adminUpdateStoreStatus
      tags: [admin]
      summary: Update store status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateStoreStatusRequest"
      responses:
        "200":
          description: Store status updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Store"
  /admin/stores/{id}/featured:
    put:
      operationId: adminToggleStoreFeatured
      tags: [admin]
      summary: Toggle store featured status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Featured status toggled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Store"
  /admin/users:
    get:
      operationId: adminListUsers
      tags: [admin]
      summary: List all users (admin)
      parameters:
        - name: page
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: All users
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserListResponse"
  /admin/users/{id}/role:
    put:
      operationId: adminUpdateUserRole
      tags: [admin]
      summary: Update user role
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserRoleRequest"
      responses:
        "200":
          description: Role updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserPublic"
  /admin/users/{id}/block:
    put:
      operationId: adminToggleUserBlock
      tags: [admin]
      summary: Toggle user block status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Block status toggled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserPublic"
  /admin/categories:
    get:
      operationId: adminListCategories
      tags: [admin]
      summary: List all categories (admin)
      responses:
        "200":
          description: All categories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Category"

    post:
      operationId: adminCreateCategory
      tags: [admin]
      summary: Create a category
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCategoryRequest"
      responses:
        "200":
          description: Category created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
  /admin/categories/{id}:
    put:
      operationId: adminUpdateCategory
      tags: [admin]
      summary: Update a category
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCategoryRequest"
      responses:
        "200":
          description: Category updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
  /admin/banners:
    get:
      operationId: adminListBanners
      tags: [admin]
      summary: List all banners (admin)
      responses:
        "200":
          description: All banners
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Banner"

    post:
      operationId: adminCreateBanner
      tags: [admin]
      summary: Create a banner
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBannerRequest"
      responses:
        "200":
          description: Banner created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Banner"
  /admin/banners/{id}:
    put:
      operationId: adminUpdateBanner
      tags: [admin]
      summary: Update a banner
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBannerRequest"
      responses:
        "200":
          description: Banner updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Banner"

    delete:
      operationId: adminDeleteBanner
      tags: [admin]
      summary: Delete a banner
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Banner deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success
...
    UserPublic:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          type: string
        avatarUrl:
          type: string
          nullable: true
        phone:
          type: string
          nullable: true
        district:
          type: string
          nullable: true
        isBlocked:
          type: boolean
        createdAt:
          type: string
          format: date-time
          nullable: true
      required:
        - id
        - name
        - email
        - role
        - isBlocked
...
    UserListResponse:
      type: object
      properties:
        users:
          type: array
          items:
            $ref: "#/components/schemas/UserPublic"
        total:
          type: integer
        page:
          type: integer
        totalPages:
          type: integer
      required:
        - users
        - total
        - page
        - totalPages
    Category:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        slug:
          type: string
        icon:
          type: string
          nullable: true
        description:
          type: string
          nullable: true
        parentId:
          type: integer
          nullable: true
        isActive:
          type: boolean
          nullable: true
        sortOrder:
          type: integer
          nullable: true
        createdAt:
          type: string
          format: date-time
          nullable: true
      required:
        - id
        - name
        - slug
    CreateCategoryRequest:
      type: object
      properties:
        name:
          type: string
        slug:
          type: string
        icon:
          type: string
        description:
          type: string
        isActive:
          type: boolean
        sortOrder:
          type: integer
      required:
        - name
        - slug
    Store:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        name:
          type: string
        slug:
          type: string
        description:
          type: string
          nullable: true
        logoUrl:
          type: string
          nullable: true
        bannerUrl:
          type: string
          nullable: true
        categoryId:
          type: integer
          nullable: true
        location:
          type: string
          nullable: true
        district:
          type: string
          nullable: true
        lat:
          type: string
          nullable: true
        lng:
          type: string
          nullable: true
        whatsapp:
          type: string
          nullable: true
        instagram:
          type: string
          nullable: true
        facebook:
          type: string
          nullable: true
        website:
          type: string
          nullable: true
        status:
          type: string
        isFeatured:
          type: boolean
          nullable: true
        totalVisits:
          type: integer
          nullable: true
        createdAt:
          type: string
          format: date-time
          nullable: true
        updatedAt:
          type: string
          format: date-time
          nullable: true
        category:
          $ref: "#/components/schemas/Category"
          nullable: true
      required:
        - id
        - userId
        - name
        - slug
        - status
...
    StoreListResponse:
      type: object
      properties:
        stores:
          type: array
          items:
            $ref: "#/components/schemas/Store"
        total:
          type: integer
        page:
          type: integer
        totalPages:
          type: integer
      required:
        - stores
        - total
        - page
        - totalPages
    CreateStoreRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        categoryId:
          type: integer
        location:
          type: string
        district:
          type: string
        lat:
          type: string
        lng:
          type: string
        whatsapp:
          type: string
        instagram:
          type: string
        facebook:
          type: string
        website:
          type: string
        logoUrl:
          type: string
        logoPublicId:
          type: string
        bannerUrl:
          type: string
        bannerPublicId:
          type: string
      required:
        - name
...
    UpdateStoreStatusRequest:
      type: object
      properties:
        status:
          type: string
          enum: [pending, active, rejected]
      required:
        - status
...
    Banner:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
          nullable: true
        subtitle:
          type: string
          nullable: true
        imageUrl:
          type: string
        publicId:
          type: string
          nullable: true
        linkUrl:
          type: string
          nullable: true
        isActive:
          type: boolean
          nullable: true
        sortOrder:
          type: integer
          nullable: true
      required:
        - id
        - imageUrl
    CreateBannerRequest:
      type: object
      properties:
        title:
          type: string
        subtitle:
          type: string
        imageUrl:
          type: string
        publicId:
          type: string
        linkUrl:
          type: string
        isActive:
          type: boolean
        sortOrder:
          type: integer
      required:
        - imageUrl
...
    UpdateUserRoleRequest:
      type: object
      properties:
        role:
          type: string
          enum: [user, vendor, admin]
      required:
        - role