# lib/api-zod/src/generated/types/storeListResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Mercanto — Marketplace Vecinal API
 * OpenAPI spec version: 0.1.0
 */
import type { Store } from "./store";

export interface StoreListResponse {
  stores: Store[];
  total: number;
  page: number;
  totalPages: number;
}